requirements.txt
alembic==1.14.1
amqp==5.3.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Deprecated==1.2.18
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.2.0
gunicorn==23.0.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
limits==4.0.1
Mako==1.3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyaes==1.6.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
sympy==1.13.1
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
